test(schemas): split generated contract checks
